pyproject.toml
[project]
name = "finsight"
version = "0.1.0"
description = ""
authors = [
    {name = "junsu0302",email = "junsu96378224@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2 (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pykrx (>=1.0.51,<2.0.0)",
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
